Rename ListStore to ListsStore
