+++ components/ShoppingListEntryItem.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ThemedText } from "./ThemedText";
-import { useShoppingListEntryCell } from "@/stores/ShoppingListStore";
+import { useShoppingListEntryCell } from "@/stores/ShoppingListsStore";
 
 export default function ShoppingListEntryItem({ id }: { id: string }) {
   const name = useShoppingListEntryCell(id, "name");
